requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# LLM
anthropic==0.43.0

# Vector Database
chromadb==0.6.3
sentence-transformers==3.4.1

# Document Parsing
PyMuPDF==1.25.3
python-docx==1.1.2

# Database
aiosqlite==0.20.0

# Frontend
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
tiktoken==0.8.0
fastapi
pydantic
pydantic-settings
python-multipart
uvicorn
aiosqlite
anthropic
python-dotenv
tenacity
tiktoken
chromadb
pydantic-settings
PyMuPDF
python-docx
python-multipart
sentence-transformers
aiosqlite
anthropic
chromadb
fastapi
pydantic
pydantic-settings
PyMuPDF
python-docx
python-dotenv
python-multipart
sentence-transformers
tenacity
tiktoken
uvicorn[standard]
plotly
